refactor(runtime): fold runtime host into agent-framework (RUNTIME-001 Design C)
Independent reviews (proposal-review + architecture-auditor) rejected a new package
(Design B) for Design C: the neutral runtime host belongs in the agent-framework
assembly layer that already owns InteractiveSession — same deps, no new package/cycle.
Owner directive (common-mistakes #80): reconcile the pre-existing duplication, do not
preserve legacy.

- Move startRuntimeHost + buildRuntimeSession (+ IRuntimeHostOptions/Handle) into
  agent-framework/src/runtime/ (next to createAgentRuntime); export them. Delete the new
  packages/agent-runtime. agent-cli serve-mode imports from agent-framework.
- Unify session construction on ONE seam: the TUI (TuiInteractionChannel), print
  (HeadlessInteractionChannel), and --serve all build via buildRuntimeSession instead of
  private `new InteractiveSession`. createAgentRuntime (lossy headless subset, only
  starter-nextjs) and createInteractiveRuntime left for their consumers; not extended.
- 5 test mocks updated: mocking InteractiveSession alone no longer intercepts
  construction (it flows through buildRuntimeSession), so mock both.
- common-mistakes #80: legacy is not the baseline; correct wrong layering / reconcile
  duplication rather than bolt new code alongside the flaw.

Verified: typecheck clean; tests framework 1088 + transport 45 + transport-tui 418 + cli
203; agent-app 12 + headless e2e 5/5; harness:scan 49/49.
