+++ .agents/rules/common-mistakes.md
@@ -84,3 +84,4 @@ Parent: [AGENTS.md](../../AGENTS.md) | Index: [rules/index.md](index.md)
 | 77  | Serializing CI (or a local runner) behind a slow upstream job/step whose output the work does not consume, and running independent check suites one-at-a-time                                                                                         | Two neutral levers: (a) a job must only `needs:` an upstream job for the steps that actually consume its artifact — split checks that don't need the artifact into a job with **no `needs`** so they run in parallel (and register that new job as a **required status check**, or a red parallel job can't block merge — the INFRA-026 gate-blind-spot); (b) independent, read-only subprocesses should run under a **bounded concurrency pool** (cap ≈ `availableParallelism()-1`), not a sequential `for..await`, capturing each child's output and printing it **only on failure** so parallel logs don't interleave while the summary/exit contract stays byte-identical. Incident: the CI `quality` job `needs: build` and its 48-scan runner (`run-all-scans.mjs`) ran subprocesses sequentially — a ~4m gate; only ONE scan (`build-contracts`) actually read `dist`, so 47 scans + the harness tests were carved into a no-`needs` `scans` job and the runner parallelized (9.9s→4.65s local) (INFRA-037, 2026-07-09).                                                                                                                                                                                                                                                                                                          |
 | 78  | For a NEW surface (package/app/UI), building it as a thin skin on an unrelated sibling PRODUCT instead of mirroring the analogous existing layer — and burying that placement decision instead of surfacing it as the primary owner call              | The architecture-**placement** of a new surface is the single most owner-critical, hardest-to-reverse decision — treat it as such (see `spec-workflow.md` "New-Surface Architecture Placement"). (a) **Mirror the analogous existing layer**: find the closest structural analog that plays the same role and justify the new surface by mirroring that proven layering (or justify differing); state its product-family/taxonomy. (b) **Reuse at the shared-core/contract level, never depend on a sibling product** that merely renders something similar — coupling a new surface under an unrelated product's package is the anti-pattern. (c) **Validate with an independent architecture review** (`proposal-reviewer`/`architecture-auditor`), not a bare "reviewed" claim; record the verdict. (d) **Surface the placement to the owner FIRST**, above styling/scope. Incident: a new desktop GUI app was proposed as a thin consumer of the browser-remote product (`agent-web-ui`) instead of mirroring the TUI layering (`agent-transport-tui` ← `agent-cli` → the GUI analog `agent-transport-gui` ← `apps/agent-app`); a proposal-review had endorsed it on narrow engineering grounds without checking layer-mirroring/product-family, and the owner caught the mis-placement only after it shipped (GUI-005, 2026-07-12). |
 | 79  | Framing one presentation surface as **controlling / owning** a sibling presentation surface — e.g. describing a desktop GUI app as "spawning and controlling the CLI", when what it actually needs is the shared **runtime** the CLI happens to embed | Presentation surfaces (TUI, GUI-app, GUI-web) are **siblings over a shared runtime/session engine — none controls another.** A GUI does not control the CLI; both drive the same runtime. When one surface reuses another's binary/process only because the runtime isn't yet extracted (e.g. a desktop app spawns the `robota` CLI in headless mode to get the runtime), (a) **frame it as "drives the shared runtime", never "controls the CLI"** in specs/docs/diagrams; (b) treat runtime reuse-via-a-sibling-binary as a pragmatic bridge, and record the clean end-state (extract a headless runtime surface both sit over) as an explicit backlog item; (c) when packaging/bundling that runtime, bundle a **headless runtime entry**, not the sibling's full product (which drags in its presentation code). Incident: `apps/agent-app` was described as spawning/controlling the `agent-cli` process; the owner corrected that the GUI does not control the CLI — both are sibling presentations over a shared runtime (GUI-002/006, 2026-07-12).                                                                                                                                                                                                                                                                               |
+| 80  | Treating the **existing / legacy structure as the standard** — placing new code where current code happens to sit, or preserving a wrong layering / duplication just because it is already there ("it's next to the existing X, so put it there too") | **Legacy is not the baseline; only correct architecture is.** That code already exists somewhere, or that a construct is already in a package, is NOT evidence it belongs there — do not treat the current structure as authoritative. When adding a surface, decide its home from correct layering/cohesion, and when you touch a layer that is already mis-structured or duplicated, **correct it (reconcile/relocate/collapse the duplication) rather than bolt the new thing alongside the flaw.** Do not preserve legacy for its own sake; do not let "minimize churn against what exists" override "put it in the right place." If an existing construct is the wrong shape, fixing it is in scope, not a separate someday. Incident: RUNTIME-001 — a runtime host was first placed in the product shell (agent-cli), then in a new package; the correct home was the existing assembly layer (agent-framework), where a sibling `createAgentRuntime` already modeled the concept — and the owner directed that the pre-existing duplication be reconciled, not preserved, on the way in (2026-07-12).                                                                                                                                                                                                                             |
