+++ packages/agent-framework/docs/SPEC.md
@@ -141,6 +141,8 @@ Core classes and functions exported from `@robota-sdk/agent-framework`:
 | `createQuery`                               | function | Single-shot prompt factory (`({ provider }) => (prompt) => result`)                                                                                                            |
 | `createAgentRuntime`                        | function | Headless/multi-session runtime composition factory                                                                                                                             |
 | `createStatelessRuntime`                    | function | Filesystem-free runtime for serverless/embedded contexts (no session store, no-op settings, bare sessions by default)                                                          |
+| `buildRuntimeSession`                       | function | RUNTIME-001: the single session-construction seam — builds an `InteractiveSession` from resolved `TInteractiveSessionOptions` (used by the TUI, print, and `--serve`)          |
+| `startRuntimeHost`                          | function | RUNTIME-001: presentation-free runtime host — builds the session + owns the transport `startAll/stopAll` + bounded shutdown handle (used by the headless `robota --serve`)     |
 | `createProjectSessionStore`                 | function | Project-local session store facade                                                                                                                                             |
 | `createUserSessionStore`                    | function | User-level session store facade (`~/.robota/sessions`)                                                                                                                         |
 | `listResumableSessionSummaries`             | function | List saved sessions for session picker UI                                                                                                                                      |
