+++ packages/agent-framework/src/runtime/index.ts
@@ -6,3 +6,5 @@ export type {
 export { createAgentRuntime } from './agent-runtime.js';
 export type { IStatelessRuntimeConfig } from './stateless-runtime.js';
 export { createStatelessRuntime } from './stateless-runtime.js';
+export { buildRuntimeSession, startRuntimeHost } from './runtime-host.js';
+export type { IRuntimeHostOptions, IRuntimeHostHandle } from './runtime-host.js';
