+++ packages/agent-transport-tui/src/__tests__/TuiInteractionChannel.lifecycle.test.ts
@@ -10,9 +10,10 @@ vi.mock('@robota-sdk/agent-framework', async () => {
   const actual = await vi.importActual<typeof import('@robota-sdk/agent-framework')>(
     '@robota-sdk/agent-framework',
   );
-  return {
-    ...actual,
-    InteractiveSession: vi.fn().mockImplementation(() => {
+  // RUNTIME-001: construction now flows through `buildRuntimeSession` (which wraps InteractiveSession inside
+  // agent-framework), so mocking `InteractiveSession` alone no longer intercepts it — mock both with one factory.
+  const makeMockSession = (): unknown => {
+    {
       const handlers = new Map<string, Array<(...args: unknown[]) => void>>();
       return {
         getFullHistory: vi.fn().mockReturnValue([]),
@@ -50,7 +51,12 @@ vi.mock('@robota-sdk/agent-framework', async () => {
         sendAgentJob: vi.fn().mockResolvedValue(undefined),
         readExecutionWorkspaceDetail: vi.fn().mockResolvedValue({}),
       };
-    }),
+    }
+  };
+  return {
+    ...actual,
+    InteractiveSession: vi.fn().mockImplementation(makeMockSession),
+    buildRuntimeSession: vi.fn().mockImplementation(makeMockSession),
     CommandRegistry: vi.fn().mockImplementation(() => ({
       addModule: vi.fn(),
     })),
